♻️ refactor: 그룹 파일 통합
